requirements.txt
# AWS SDK
boto3>=1.34.0

# Environment management
python-dotenv>=1.0.0

# Datadog tracing and observability
ddtrace>=1.20.0

# Type hints support (for Python < 3.9)
typing-extensions>=4.0.0

# JSON handling
jsonschema>=4.0.0

# Logging and utilities
colorama>=0.4.0 